+++ src/main/webapp/components/gallery/index.js
